change files structure
