+++ app.py
@@ -1,4 +1,5 @@
-from muse_as_service import create_app, download_thhub_model, get_argparse
+from muse_as_service.flask_app import create_app
+from muse_as_service.utils import download_thhub_model, get_argparse
 
 app = create_app()
