+++ muse_as_service/utils.py
@@ -102,3 +102,24 @@ def download_thhub_tar_model(
         # save compressed model
         with open(save_model_path + ".tar", mode="wb") as fp:
             fp.write(response.content)
+
+
+def download_thhub_model(
+    thhub_model_url: str = "https://tfhub.dev/google/universal-sentence-encoder-multilingual/3",
+    save_model_path: str = ".cache/universal-sentence-encoder-multilingual_3",
+    verbose: bool = True,
+) -> None:
+    """
+    Download tf hub .tar model given URL and unpack it.
+
+    :param str thhub_model_url: tf hub model URL.
+    :param str save_model_path: path to save model.
+    :param bool verbose: verbose.
+    """
+
+    download_thhub_tar_model(
+        thhub_model_url=thhub_model_url,
+        save_model_path=save_model_path,
+        verbose=verbose,
+    )
+    unpack_tar(path=save_model_path + ".tar", remove=True, verbose=verbose)  # hardcode
